Module picky_asn1::wrapper

Structs

Asn1SequenceOf

Asn1 wrapper around a collection of elements of the same type.

Asn1SetOf

Asn1 wrapper around a collection of elements of the same type.

BMPStringAsn1

Wrapper type

BitStringAsn1

Wrapper type

BitStringAsn1Container

A BitString encapsulating things.

ExplicitContextTag0

Wrapper type for special tag

ExplicitContextTag1

Wrapper type for special tag

ExplicitContextTag2

Wrapper type for special tag

ExplicitContextTag3

Wrapper type for special tag

ExplicitContextTag4

Wrapper type for special tag

ExplicitContextTag5

Wrapper type for special tag

ExplicitContextTag6

Wrapper type for special tag

ExplicitContextTag7

Wrapper type for special tag

ExplicitContextTag8

Wrapper type for special tag

ExplicitContextTag9

Wrapper type for special tag

ExplicitContextTag10

Wrapper type for special tag

ExplicitContextTag11

Wrapper type for special tag

ExplicitContextTag12

Wrapper type for special tag

ExplicitContextTag13

Wrapper type for special tag

ExplicitContextTag14

Wrapper type for special tag

ExplicitContextTag15

Wrapper type for special tag

GeneralStringAsn1

Wrapper type

GeneralizedTimeAsn1

Wrapper type

HeaderOnly

A wrapper encoding/decoding only the header of the provided Asn1Wrapper with a length of 0.

IA5StringAsn1

Wrapper type

ImplicitContextTag0

Wrapper type for special tag

ImplicitContextTag1

Wrapper type for special tag

ImplicitContextTag2

Wrapper type for special tag

ImplicitContextTag3

Wrapper type for special tag

ImplicitContextTag4

Wrapper type for special tag

ImplicitContextTag5

Wrapper type for special tag

ImplicitContextTag6

Wrapper type for special tag

ImplicitContextTag7

Wrapper type for special tag

ImplicitContextTag8

Wrapper type for special tag

ImplicitContextTag9

Wrapper type for special tag

ImplicitContextTag10

Wrapper type for special tag

ImplicitContextTag11

Wrapper type for special tag

ImplicitContextTag12

Wrapper type for special tag

ImplicitContextTag13

Wrapper type for special tag

ImplicitContextTag14

Wrapper type for special tag

ImplicitContextTag15

Wrapper type for special tag

IntegerAsn1

A BigInt wrapper for Asn1 encoding.

NumericStringAsn1

Wrapper type

ObjectIdentifierAsn1

Wrapper type

OctetStringAsn1

A Vec wrapper for Asn1 encoding as OctetString.

OctetStringAsn1Container

An OctetString encapsulating things.

Optional

Wrapper for ASN.1 optionals fields

PrintableStringAsn1

Wrapper type

UTCTimeAsn1

Wrapper type

Utf8StringAsn1

Wrapper type